Split tests to mirror module layout (shell / contrib / registry)
`tests/test_contrib.py` covers the best-effort shells (fish, csh/tcsh,
xonsh). `tests/test_registry.py` covers `SHELLS`, `default_shell_class`,
and `detect_shell_class`. `tests/test_shell.py` keeps tests for the
fully-supported shells, the hook/integration tests, and the abstract
`UnixShell` check.

Move the session-scoped `simple_env` fixture to `conftest.py` so all
three test modules can share it.
